nodesCoordinator: moved structs from interface file to dtos file
