+++ sharding/nodesCoordinator/interface.go
@@ -49,26 +49,6 @@ type PublicKeysSelector interface {
 	GetOwnPublicKey() []byte
 }
 
-// ArgsUpdateNodes holds the parameters required by the shuffler to generate a new nodes configuration
-type ArgsUpdateNodes struct {
-	Eligible          map[uint32][]Validator
-	Waiting           map[uint32][]Validator
-	NewNodes          []Validator
-	UnStakeLeaving    []Validator
-	AdditionalLeaving []Validator
-	Rand              []byte
-	NbShards          uint32
-	Epoch             uint32
-}
-
-// ResUpdateNodes holds the result of the UpdateNodes method
-type ResUpdateNodes struct {
-	Eligible       map[uint32][]Validator
-	Waiting        map[uint32][]Validator
-	Leaving        []Validator
-	StillRemaining []Validator
-}
-
 // NodesShuffler provides shuffling functionality for nodes
 type NodesShuffler interface {
 	UpdateParams(numNodesShard uint32, numNodesMeta uint32, hysteresis float32, adaptivity bool)
